requirements.txt
python-telegram-bot
pyshorteners